requirements.txt
aiofiles==25.1.0
aiohttp==3.13.3
alembic==1.18.0
beautifulsoup4==4.14.2
botbuilder-core==4.17.1
celery[redis]==5.6.2
email-validator==2.3.0
fastapi[standard-no-fastapi-cloud-cli]
google-adk==1.22.0
httpx==0.28.1
instructor==1.14.1
litellm==1.80.13
markdown-pdf==1.10
psycopg[binary,pool]==3.3.2
pwdlib[argon2] >=0.3.0
pydantic_settings==2.12.0
PyGithub==2.8.1
pyjwt==2.10.1
PyMuPDF>=1.25.3
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
qdrant-client==1.16.2
slowapi==0.1.9
SQLAlchemy==2.0.45
